feture: Add error 500 and Errors components functionality and .js and css files
