+++ src/Components/ErrorHandling/Errors.js
@@ -1,16 +1,36 @@
-import React from 'react';
-import { NavLink } from 'react-router-dom';
+import React, { useEffect } from 'react';
+import { NavLink, useNavigate } from 'react-router-dom';
 import PropTypes from 'prop-types';
-// import Groundhog from '../../images/groundhog.png';
+import GreatMinds from '../../Images/greatminds.jpg'
 import './Errors.css';
 
-function Errors({ error }) {
+function Errors({ error, setError, setTriviaLoading }) {
+  const navigate = useNavigate();
+
+  function handleError(error) {
+    if (error.message === '500') {
+      setError('500');
+    } else if (error.message === '404') {
+      setError('404');
+    } else {
+      setError('Other');
+    }
+    setTriviaLoading(false);
+    navigate('/error');
+  }
+
+  useEffect(() => {
+    if (error) {
+      handleError(error);
+    }
+  }, [error]);
+
   return (
     <div className="error-msg">
       <h1>Oops! We're looking for it!</h1>
-      {/* <img src={Groundhog} alt="Groundhog" /> */}
+      <img src={GreatMinds} alt="Grait Mind of 980 EC" />
       <NavLink to="/" style={{ textDecoration: 'none' }}>
-        <button className="go-home-button" aria-label="Go back to the home page">
+        <button className="errors-go-home-button" aria-label="Go back to the home page">
           Go Back to Home Page
         </button>
       </NavLink>
@@ -31,6 +51,9 @@ Errors.propTypes = {
     message: PropTypes.string,
     statusCode: PropTypes.number,
   }),
+  setError: PropTypes.func.isRequired,
+  setTriviaLoading: PropTypes.func.isRequired,
 };
+
 
 export default Errors;
