+++ src/Components/Flashcard/Flashcard.css
@@ -22,10 +22,10 @@ body {
 
 .form-group select,
 .form-group input[type="number"] {
-  width: 100%; 
+  width: 100%;
   font-size: 1.2rem;
   padding: 0.5rem;
-  width: 100%; 
+  width: 100%;
   margin-right: 2rem;
   margin-top: 0.78rem;
 }
@@ -64,7 +64,7 @@ body {
 .btn:hover {
   background-color: rgb(13, 246, 5);
   box-shadow: 4px 4px 8px rgba(0, 0, 0, 0.3);
-  transform: translateY(-2px); 
+  transform: translateY(-2px);
 }
 
 .button.go-back-button {
@@ -82,9 +82,9 @@ body {
 }
 
 .button.go-back-button:hover {
-   background-color: hsl(200, 100%, 50%);
-   box-shadow: 4px 4px 8px rgba(0, 0, 0, 0.3);
-   transform: translateY(-2px); 
+  background-color: hsl(200, 100%, 50%);
+  box-shadow: 4px 4px 8px rgba(0, 0, 0, 0.3);
+  transform: translateY(-2px);
 }
 
 .card-grid {
@@ -156,8 +156,8 @@ body {
 
 .correct-answer-emoji {
   position: absolute;
-  top: 0.1rem; 
-  right: 8rem; 
+  top: 0.1rem;
+  right: 8rem;
 }
 
 img {
@@ -172,13 +172,13 @@ img {
   justify-content: center;
   align-items: center;
   height: 100vh;
-  background-color:white;
+  background-color: white;
 }
 
 .spinner {
   border: 4px solid hsl(200, 100%, 50%);
   width: 92px;
-  height: 92px; 
+  height: 92px;
   border-radius: 50%;
   border-left-color: #E50914;
   animation: spin 1s linear infinite;
@@ -188,6 +188,7 @@ img {
   0% {
     transform: rotate(0deg);
   }
+
   100% {
     transform: rotate(360deg);
   }
